+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 * doc(filteredNavigation): improve
 * feat(plugins\platform): detect `bitrixMobile`
-* feat(Theme): new component
 * fix(Prose.A): add prop `raw`
+* feat(Theme): new component
+* fix(ColorModeImage): add baseURL support for public paths
 
 ## [2.3.0](https://github.com/bitrix24/b24ui/compare/v2.2.1...v2.3.0) (2026-02-12)